Update patch set 2
Patch Set 2:

The base image is per https://phabricator.wikimedia.org/T205911#4982711 unfortunately a dead end. With all the ways that nodejs would discover the 2 packages being deprecated and/or discouraged following that path seems unwise (and would anyway require even more changes into blubber to support it), hence this.

We can always move the 2 packages in the blubber.yaml files at the end of the day and do a lot of copy/pasting but the entirety of T205911 is about avoiding doing that. Since we have policies, would it make sense to add support to policies to augment the list of packages for nodejs apps and use that policy to do so?

Patch-set: 2
Reviewer: Alexandros Kosiaris <824@e9e9afe9-4712-486d-8885-f54b72dd1951>