Update github.com/docker/docker dependency
Updated to github.com/docker/docker v26.1.3+incompatible.

Unfortunately some transitive dependency breakage occurred with the
update. The root cause of the issue is described at
https://github.com/docker/cli/issues/4812#issuecomment-1905732686

> This repository is not yet a go modules (there's no go.mod), which
> currently means that go mod is not able to resolve / detect versions of
> (indirect) dependencies that are needed. [...] unfortunately OTEL is
> notorious for being difficult in this area (there's many inter-connected
> libraries, some of which are incompatible if not at the exact version
> needed, and not all of them have "correct" requires rules defined).

To work around the problem, a number of replace directives have been
added to go.mod that pin the versions of otel related dependencies.

Incidentally, the similar hack that was in place for the logrus
dependency is no longer an issue and has been removed.

Change-Id: I914513bfef3548e7b79f4f9e2719cb03b1b7fa40