build: Reset image default arguments (i.e. CMD) when entrypoint is defined
Dockerfile behavior, and by extension that of the prior Dockerfile based
blubber, resets the CMD value inherited by base images when an
ENTRYPOINT is defined. To maintain parity (for now at least) with the
prior implementation, have `build.TargetImage` reset the image's default
arguments when a new entry point is set.

Change-Id: I35bf0e160067568597863852857fc2fb539b6c94