apt: Support `signed-by` field in apt configuration
Support verification of additional APT sources via a new `signed-by`
field. The value is expected to be the ASCII encoded public key for the
given source.

Provide a new `build.File` instruction for creation of new files with
literal byte content and refactor prior use of `echo >>` commands to use
the new instruction.

Write the contents of `signed-by` prior to the setup of each additional
APT source.

Install `ca-certificates` as a bootstrap measure prior to adding sources
to ensure that fetching of https sources succeed.

Bug: T395645
Change-Id: Id220dc4c6071e50fdc5e7eb8020d99d5e67eb289