builder: Support multiline builder `script` and string `command`
Support configuration of multiline builder scripts via a new `script`
field. The script is not present in the resulting image, but rather its
content is written to a temporary scratch filesystem prior to execution
where it is mounted readonly.

Example

```
builder:
  script: |
    #!/bin/bash
    if ! [[ "$CGO_ENABLED" == "0" ]]; then
      echo "you must set CGO_ENABLED=0 for this build"
      exit 1
    fi
    go build $(go list)
```

In addition, support a string form for `command`. This is much less
awkward than the JSON array form. Refactor all builder examples to use
the string form.

The `script` and `command` fields are mutually exclusive, validated via
`notallowedwith`. Refactored the `notallowedwith` validator to reference
other fields by JSON name so that errors are more intuitive to end
users.

Allow builders with no command. This let's folks avoid the weird
`command: ["true"]` loophole.

Change-Id: I8bc650faadf5090ed87546ca2f9a2878f8e6c9f5