builder: Support multiline builder `script` and string `command`
Support configuration of multiline builder scripts via a new `script`
field. The script is not present in the resulting image, but rather its
content is written to a temporary scratch filesystem prior to execution
where it is mounted readonly.

Example

```
builder:
  script: |
    #!/bin/bash
    if ! [[ "$CGO_ENABLED" == "0" ]]; then
      echo "you must set CGO_ENABLED=0 for this build"
      exit 1
    fi
    go build $(go list)
```

In addition, support a string form for `command`. This is much less
awkward than the JSON array form. Refactor all builder examples to use
the string form.

Allow builders with no command. This let's folks avoid the weird
`command: ["true"]` loophole.

Change-Id: Id31fab89a50ba4fb5be6d5cdbb62f74df3cf884d