mounts: New read/write behavior and shorthand configuration
Read-only mounts can cause issues for builders that require read/write
access to source directories. Make all source mounts read/write and
mention that changes to files under the mount point are discarded after
each builder process finishes (the behavior of all source mounts in
BuildKit).

Default the `destination` path of source mounts to the working directory
"." and refactor the shorthand configuration to reference the `from`
instead of the `destination` path of the local context. The latter is of
marginal utility while the former can be used to reference any named
context that should be mounted as the working directory. This is a
breaking change but the `mounts` feature is brand new and not in use
according to code search.

Change-Id: I0ad9fb974617b959e43ac84853c00c892b17e857