ci: Remove `BUILD_CONTEXT` from bake file
Explicitly setting the target `context` from the `BUILD_CONTEXT`
variable was meant to workaround an issue where the `.git` directory
could not be kept when the default bake context was relied on. However,
this issue does not arise if you explicitly specify `context = "."` (the
remote bake context _path_ is correctly used but the context _state_ is
not reused).

See https://github.com/docker/buildx/pull/3338 for more information and
a fix for the case where no target `context` is defined.

Change-Id: I004f1d558e3f8cf0a0f48d1a72f6f9644c1a5070