docs: Refactor README to focus on common usage patterns
Rewrite the `README.md` to focus on the most common usage patterns and
suggest `docker buildx build` over `buildctl` as the latter is meant for
internal and development use according to upstream.

Include sections for usage with Docker Bake and Docker Compose.

Consolidate sections on build arguments, multiplatform support, and
attestations by linking to relevant sections of upstream documentation.

Link to examples and configuration reference sections at
https://doc.wikimedia.org/releng/blubber instead of the local markdown
and feature files.

Change-Id: I901bab39f08f76bd22c78a9c11ee20b24a44ea91