buildkit: Support `--no-cache` client options
Support `--no-cache` client options in the Blubber frontend gateway.

Define `NoCache` build option as a function that takes a target name and
returns whether the target's instructions ignore the cache.

Set the `NoCache` option to `dockerui.Client.IsNoCache` to ensure the
gateway is compatible with `docker build` and `docker buildx` CLIs.

Change-Id: I2218761d10f0f926b55e34e86140e6b00a9f6ce3